add: release: support for socials, now can be shared from company
